+++ src/components/button/button.test.ts
@@ -4,7 +4,7 @@ import { runFormControlBaseTests } from '../../internal/test/form-control-base-t
 import sinon from 'sinon';
 import type SlButton from './button.js';
 
-const variants = ['default', 'primary', 'success', 'neutral', 'warning', 'danger'];
+const variants = ['brand', 'success', 'neutral', 'warning', 'danger'];
 
 describe('<sl-button>', () => {
   describe('accessibility tests', () => {
@@ -33,7 +33,6 @@ describe('<sl-button>', () => {
       expect(el.loading).to.equal(false);
       expect(el.outline).to.equal(false);
       expect(el.pill).to.equal(false);
-      expect(el.circle).to.equal(false);
     });
 
     it('should render as a <button>', async () => {