+++ src/components/tag/tag.component.ts
@@ -31,7 +31,7 @@ export default class SlTag extends ShoelaceElement {
   private readonly localize = new LocalizeController(this);
 
   /** The tag's theme variant. */
-  @property({ reflect: true }) variant: 'primary' | 'success' | 'neutral' | 'warning' | 'danger' | 'text' = 'neutral';
+  @property({ reflect: true }) variant: 'brand' | 'success' | 'neutral' | 'warning' | 'danger' | 'text' = 'neutral';
 
   /** The tag's size. */
   @property({ reflect: true }) size: 'small' | 'medium' | 'large' = 'medium';
@@ -54,7 +54,7 @@ export default class SlTag extends ShoelaceElement {
           tag: true,
 
           // Types
-          'tag--primary': this.variant === 'primary',
+          'tag--brand': this.variant === 'brand',
           'tag--success': this.variant === 'success',
           'tag--neutral': this.variant === 'neutral',
           'tag--warning': this.variant === 'warning',